例外処理をしっかりと&Makefile設置その他諸々
